chore: move functions, lint fix
